\name{identify.point}
\title{Identify points on a Point Object}
\usage{identify(point.obj, v) }
\alias{identify.point}
\arguments{
\item{point.obj}{a point object generated by \code{point()}}
\item{v}{a variable name that will be displayed on the graphics device}
}
\value{NULL}
\description{Plot variable values next to locations after the \code{plot.point()} function. }
\seealso{\code{\link{plot.point}}}
\examples{
\testonly{
# prepare variables from other example pages if they are not already there:
if(length(ls(pat="maas.point"))==0){example(point)}
}
plot(maas.point)
# use indices as labels:
identify(maas.point)
# use values as labels:
identify(maas.point,"zinc")
}
\references{
http://www.gis.iastate.edu/SGeoStat/homepage.html
}
