\name{maas.bank}
\title{maas.bank - coordinates}
\usage{maas.bank}
\description{
Coordinates of maas bank. To be used together with \link{\code{maas}}.
}
\value{list with components \code{x} and\code{y}..
}
\seealso{\link{\code{maas}}
\references{
  gstat
  E.J Pebesma (E.J.Pebesma@frw.uva.nl) 
  http://www.frw.uva.nl/~pebesma/gstat/
}
\keyword{data}