\name{plot.variogram}
\title{Plot Variogram}
\usage{plot(variogram.obj, variogram.model.obj) }
\alias{plot.variogram}
\arguments{
\item{variogram.obj}{a variogram object generated by \code{est.variogram()}}
\item{variogram.model.obj}{a variogram model object generated by a model fitting routine.}
}
\value{NULL }
\description{Plot empirical variogram estimates, optionally plotting a fitted variogram model. }
\seealso{\code{\link{est.variogram}}}
\examples{
\testonly{
# prepare variables from other example pages if they are not already there:
if(length(ls(pat="maas.vmod"))==0){example(fit.variogram)}
}
# two plots 
oldpar <- par(mfrow=c(2,1))
plot(maas.v) 
plot(maas.v,maas.vmod)
par(oldpar)
}
\references{
http://www.gis.iastate.edu/SGeoStat/homepage.html
}
