\name{print.pair}
\title{Pairs Object Description}
\usage{print.pair(pair.obj) }
\alias{print.pair}
\arguments{
\item{pairs.obj}{a pair object generated by \code{pair()}}
}
\value{NULL}
\description{Print descriptive information about a pair object. }
\seealso{\code{\link{pair}}}
\examples{
\testonly{
# prepare variables from other example pages if they are not already there:
if(length(ls(pat="maas.pair"))==0){example(pair)}
}
print(maas.pair)
# gives:
# Pairs object: maas.pair 
#
#      Type:             isotropic
#      Number of pairs:  8370
#      Number of lags:   10
#      Max h:            1999.867
}
\references{
http://www.gis.iastate.edu/SGeoStat/homepage.html
}
\keyword{sgeostat}
