\name{print.point}
\title{Point Object Description}
\usage{print.point(point.obj) }
\alias{print.point}
\arguments{
\item{point.obj}{a point object generated by \code{point()}}
}
\value{NULL}
\description{Print descriptive information about a point object.}
\seealso{\code{\link{point}}}
\examples{
\testonly{
if(length(ls(pat="maas.point"))==0){example(point)}
}
print.point(maas.point)
# gives
# Point object: maas.point 
#
#   Locations: 155
#
#   Attributes:
#      x
#      y
#      zinc
}
\references{
http://www.gis.iastate.edu/SGeoStat/homepage.html
}
