\name{sgeostat-internal}
\alias{calcangle}
\alias{export.point}
\alias{krige.all}
\alias{krige.maxdist}
\alias{pair.aniso}
\alias{pair.iso}
\alias{pair.newangle}
\alias{prediction.matrix}
\alias{trend.matrix}
\alias{trend.value}
\alias{which.na}
\title{Internal sgeostat functions}
\description{
 Internal sgeostat functions
}
\usage{
The \code{\link{krige()}} function interfaces to \code{krige.*}, \code{\link{pair}}
to \code{pair.*} and \code{\link{fit.trend}} to \code{trend.*}.
}
\details{
  These are not to be called by the user.
}
\keyword{sgeostat}
\keyword{sgeostat}
