\name{maas}
\alias{maas}
\title{maas- zinc measurements}
\usage{maas}
\description{
Zinc measurements as groundwater quality variable.
}
\value{list with components \code{x},\code{y} and \code{zinc}.
}
\references{
  gstat
  E.J Pebesma (E.J.Pebesma@frw.uva.nl) 
  http://www.frw.uva.nl/~pebesma/gstat/
}
\seealso{\code{\link{maas.bank}}}
\keyword{datasets}
