\name{print.pairs}
\title{Pairs Object Description}
\usage{print.pairs(pairs.obj) }
\alias{print.pairs}
\arguments{
\item{pairs.obj}{a pairs object generated by \code{pairs()}}
}
\value{NULL}
\description{Print descriptive information about a pairs object. }
\seealso{\code{\link{pairs}}}
\examples{
\testonly{
# prepare variables from other example pages if they are not already there:
if(length(ls(pat="maas.pairs"))==0){example(pairs)}
}
print(maas.pairs)
# gives:
# Pairs object: maas.pairs 
#
#      Type:             isotropic
#      Number of pairs:  8370
#      Number of lags:   10
#      Max h:            1999.867
}
\references{
http://www.gis.iastate.edu/SGeoStat/homepage.html
}
