\name{print.point}
\title{Point Object Description}
\usage{print.point(x,...) }
\alias{print.point}
\arguments{
  \item{x}{a point object generated by \code{point()}}
  \item{...}{additional arguments for \code{print}}
}
\value{NULL}
\description{Print descriptive information about a point object.}
\seealso{\code{\link{point}}}
\examples{
\testonly{
if(length(ls(pat="maas.point"))==0){example(point)}
}
print.point(maas.point)
# gives
# Point object: maas.point 
#
#   Locations: 155
#
#   Attributes:
#      x
#      y
#      zinc
}
\references{
http://www.gis.iastate.edu/SGeoStat/homepage.html
}
\keyword{spatial}
