\name{spacebox}
\title{Boxplot of Variogram Cloud}
\usage{spacebox(point.obj, pair.obj, a1, a2, type='r')}
\alias{spacebox}
\arguments{
\item{point.obj}{a point object generated by \code{point()}}
\item{pair.obj}{a pairs object generated by \code{pair()}}
\item{a1}{a variable to plot}
\item{a2}{an optional variable name, if entered the plot will be created between \code{a1} and \code{a2}}
\item{type}{either \code{'r'} for square-root differences or \code{'s'} for squared differences}
}
\value{NULL}
\description{Create  boxplots of square-root or squared differences between variable values at pairs of points versus the distance
       between the points. }
\seealso{\code{\link{point}}, \code{\link{pair}}}
\examples{
\testonly{
# prepare variables from other example pages if they are not already there:
if(length(ls(pat="maas.pair"))==0){example(pair)}
}
spacebox(maas.point,maas.pair,'zinc')
}
\references{
http://www.gis.iastate.edu/SGeoStat/homepage.html
}
\keyword{spatial}
